requirements.txt
discord.py
sqlalchemy
requests
beautifulsoup4
snscrape
google-generativeai